Unsquashfs: make read_fs_bytes() match Mksquashfs
Read_fs_bytes() in Mksquashfs takes a long long bytes count since
a commit last May.

Read_fs_bytes() in Unsquashfs takes an int.  This should be increased
to long long to match Mksquashfs.

This is because read_xattrs.o is shared by both Mksquashfs and
Unsquashfs, and the definition for read_fs_bytes() cannot match
both implementations.

+++ squashfs-tools/read_xattrs.c
@@ -38,7 +38,7 @@
 
 #include <stdlib.h>
 
-extern int read_fs_bytes(int, long long, int, void *);
+extern int read_fs_bytes(int, long long, long long, void *);
 extern int read_block(int, long long, long long *, int, void *);
 
 static struct hash_entry {
